NP-34646 General component to serach for NFR projects
