--- src/pages/registration/description_tab/NfrProjectSearch.tsx
@@ -1,80 +1,0 @@
-import { Autocomplete, Box, Typography } from '@mui/material';
-import { Field, FieldProps } from 'formik';
-import { useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { VerifiedFundingApiPath } from '../../../api/apiPaths';
-import { AutocompleteTextField } from '../../../components/AutocompleteTextField';
-import { SearchResponse } from '../../../types/common.types';
-import { NfrProject } from '../../../types/project.types';
-import { SpecificFundingFieldNames } from '../../../types/publicationFieldNames';
-import { Funding } from '../../../types/registration.types';
-import { getPeriodString } from '../../../utils/general-helpers';
-import { useDebounce } from '../../../utils/hooks/useDebounce';
-import { useFetch } from '../../../utils/hooks/useFetch';
-import { getLanguageString } from '../../../utils/translation-helpers';
-
-interface NfrProjectSearchProps {
-  baseFieldName: string;
-}
-
-export const NfrProjectSearch = ({ baseFieldName }: NfrProjectSearchProps) => {
-  const { t } = useTranslation();
-  const [searchTerm, setSearchTerm] = useState('');
-  const debouncedSearchTerm = useDebounce(searchTerm);
-
-  const [nfrProjectSearch, isLoadingNfrProjectSearch] = useFetch<SearchResponse<NfrProject>>({
-    url: debouncedSearchTerm ? `${VerifiedFundingApiPath.Nfr}?term=${debouncedSearchTerm}&size=20` : '',
-  });
-  const projects = nfrProjectSearch?.hits ?? [];
-
-  return (
-    <Field name={`${baseFieldName}.${SpecificFundingFieldNames.Id}`}>
-      {({ field: { name, onBlur }, form: { setFieldValue }, meta: { touched, error } }: FieldProps<string>) => (
-        <Autocomplete
-          options={projects}
-          filterOptions={(options) => options}
-          onInputChange={(_, newInputValue) => setSearchTerm(newInputValue)}
-          getOptionLabel={(option) => getLanguageString(option.labels)}
-          renderOption={(props, option) => {
-            const projectName = getLanguageString(option.labels);
-            const period = getPeriodString(option.activeFrom, option.activeTo);
-            const manager = option.lead ? `${t('project.project_manager')}: ${option.lead}` : '';
-            return (
-              <li {...props}>
-                <Box sx={{ display: 'flex', flexDirection: 'column' }}>
-                  <Typography sx={{ fontWeight: 600 }}>{projectName}</Typography>
-                  <Typography variant="body1">{period}</Typography>
-                  <Typography variant="body2">{manager}</Typography>
-                </Box>
-              </li>
-            );
-          }}
-          onChange={(_, value) => {
-            if (value) {
-              const { lead, ...rest } = value;
-              const nfrFunding: Funding = {
-                type: 'ConfirmedFunding',
-                ...rest,
-              };
-              setFieldValue(baseFieldName, nfrFunding);
-            }
-          }}
-          loading={isLoadingNfrProjectSearch}
-          renderInput={(params) => (
-            <AutocompleteTextField
-              {...params}
-              name={name}
-              onBlur={onBlur}
-              label={t('registration.description.funding.nfr_project')}
-              isLoading={isLoadingNfrProjectSearch}
-              placeholder={t('registration.description.funding.nfr_project_search')}
-              showSearchIcon
-              required
-              errorMessage={touched && !!error ? error : undefined}
-            />
-          )}
-        />
-      )}
-    </Field>
-  );
-};
